backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
transformers
torch
pillow
pytest
torchvision
